requirements.txt
asgiref==3.11.0
Django==6.0.1
paho-mqtt==2.1.0
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
wheel==0.45.1
